// src/api/llm/model/chat-text-chat-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * llm
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ChatTextChatRequest {
    'autoFillHistory'?: boolean;
    'configId': number;
    'conversationId'?: string;
    'message': string;
    'systemPrompt'?: string;
}