// src/api/llm/model/chatmessage-list-chat-message-by-session200-response-messages-inner-tool-calls-inner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * llm
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ChatmessageListChatMessageBySession200ResponseMessagesInnerToolCallsInnerInfo } from './chatmessage-list-chat-message-by-session200-response-messages-inner-tool-calls-inner-info';

export interface ChatmessageListChatMessageBySession200ResponseMessagesInnerToolCallsInner {
    'error': string;
    'info': ChatmessageListChatMessageBySession200ResponseMessagesInnerToolCallsInnerInfo;
    'result': string;
    'status': string;
}